(PC-5825): Refacto ExternalLink
