+++ src/features/auth/signup/SetBirthday.tsx
@@ -12,7 +12,7 @@ import { BottomCardContentContainer } from 'ui/components/BottomCard'
 import { BottomContentPage } from 'ui/components/BottomContentPage'
 import { ButtonPrimary } from 'ui/components/buttons/ButtonPrimary'
 import { ButtonTertiary } from 'ui/components/buttons/ButtonTertiary'
-import { ExternalLink } from 'ui/components/buttons/ExternalLink'
+import { ExternalLink } from 'ui/components/buttons/externalLink/ExternalLink'
 import { DateInput } from 'ui/components/inputs/DateInput'
 import { InputError } from 'ui/components/inputs/InputError'
 import { AppInformationModal } from 'ui/components/modals/AppInformationModal'
