+++ src/features/cheatcodes/pages/AppComponents.tsx
@@ -14,7 +14,7 @@ import { ButtonQuaternary } from 'ui/components/buttons/ButtonQuaternary'
 import { ButtonSecondary } from 'ui/components/buttons/ButtonSecondary'
 import { ButtonTertiary } from 'ui/components/buttons/ButtonTertiary'
 import { ButtonTertiaryWhite } from 'ui/components/buttons/ButtonTertiaryWhite'
-import { ExternalLink } from 'ui/components/buttons/ExternalLink'
+import { ExternalLink } from 'ui/components/buttons/externalLink/ExternalLink'
 import { DateInput } from 'ui/components/inputs/DateInput'
 import { PartialDateInput, DatePartType } from 'ui/components/inputs/PartialDateInput'
 import { PasswordInput } from 'ui/components/inputs/PasswordInput'
