+++ src/ui/components/buttons/externalLink/ExternalLink.tsx
@@ -2,6 +2,7 @@ import React from 'react'
 import { Text } from 'react-native'
 
 import { openExternalUrl } from 'features/navigation/helpers'
+import { extractExternalLinkParts } from 'ui/components/buttons/externalLink/ExternalLink.service'
 import { ExternalSite } from 'ui/svg/icons/ExternalSite'
 import { ColorsEnum, getSpacing, Typo } from 'ui/theme'
 
@@ -15,12 +16,7 @@ interface Props {
 export const ExternalLink: React.FC<Props> = ({ url, text, color, testID }) => {
   text = text || url
 
-  const firstSpaceOccurrence = text.indexOf(' ')
-  const hasOccurrence = firstSpaceOccurrence > 0
-
-  const firstWord = (hasOccurrence ? '\u00a0' : '') + text.substring(0, firstSpaceOccurrence)
-  const remainingWords =
-    (hasOccurrence ? ' ' : '\u00a0') + text.substring(firstSpaceOccurrence + 1, text.length)
+  const [firstWord, remainingWords] = extractExternalLinkParts(text)
 
   return (
     <Typo.ButtonText color={color} onPress={() => openExternalUrl(url)} testID={testID}>
